Updating Reservation class and adding Room class
