+++ src/main/java/com/robin/hoteless/Main.java
@@ -85,7 +85,7 @@ public class Main {
         } while (parseError);
 
         // compute checkOut date
-        LocalDate checkOut = IncrementDate(checkIn, nights);
+        LocalDate checkOut = incrementDate(checkIn, nights);
 
         // calculate total cost of stay based on bedCost and # of nights
         double total = bedCost * nights;
@@ -122,46 +122,8 @@ public class Main {
     }
 
     // method: increment dates for check-out calculation
-    public static LocalDate IncrementDate(LocalDate initialDate, int daysToIncrement) {
+    public static LocalDate incrementDate(LocalDate initialDate, int daysToIncrement) {
         LocalDate incrementedDate = initialDate.plusDays(daysToIncrement);
         return incrementedDate;
-    }
-
-    static class Reservation {
-        int id;
-        Date made;
-        int stayDays;
-        double dailyRate;
-        LocalDate in;
-        LocalDate out;
-        String bedType;
-        double total;
-        static int totalReservations;
-
-        Reservation(Date m, int s, LocalDate i, LocalDate o, double d, String b, double t) {
-            id = totalReservations + 10000;
-            made = m;
-            stayDays = s;
-            in = i;
-            out = o;
-            dailyRate = d;
-            bedType = b;
-            total = t;
-            totalReservations++;
-        }
-
-        void show() {
-            System.out.printf(
-                    "   - Your confirmation number is %s.\n" +
-                            "   - We've reserved a %s room for %d days starting on %s and ending on %s\n" +
-                            "   - Based on your daily rate of %s, the total cost of your stay will be $%.2f \n",
-                    id, bedType, stayDays, in, out, dailyRate,total);
-
-        }
-
-        static void showReservations() {
-            System.out.println(totalReservations);
-        }
-
     }
 }
