+++ src/main/java/com/robin/hoteless/Room.java
@@ -1,0 +1,8 @@
+package com.robin.hoteless;
+
+public class Room {
+    int bedCount;
+    String bedSize;
+    Boolean includeBreakfast;
+    Boolean includeBusinessServices;
+}
